Move duplicate code to node.py
