+++ gns3/modules/virtualbox/virtualbox_vm.py
@@ -39,11 +39,13 @@ class VirtualBoxVM(Node):
     :param project: Project instance
     """
 
+    URL_PREFIX = "virtualbox"
+    NIO_URL_PREFIX = "adapters"
+
     def __init__(self, module, server, project):
 
         Node.__init__(self, module, server, project)
         log.info("VirtualBox VM instance is being created")
-        self._vm_id = None
         self._linked_clone = False
         self._export_directory = None
         self._loading = False
@@ -138,36 +140,6 @@ class VirtualBoxVM(Node):
             self.created_signal.emit(self.id())
             self._module.addNode(self)
 
-    def delete(self):
-        """
-        Deletes this VirtualBox VM instance.
-        """
-
-        log.debug("VirtualBox VM instance {} is being deleted".format(self.name()))
-        # first delete all the links attached to this node
-        self.delete_links_signal.emit()
-        if self._vm_id:
-            self.httpDelete("/virtualbox/vms/{vm_id}".format(project_id=self._project.id(), vm_id=self._vm_id),
-                            self._deleteCallback)
-        else:
-            self.deleted_signal.emit()
-            self._module.removeNode(self)
-
-    def _deleteCallback(self, result, error=False, **kwargs):
-        """
-        Callback for delete.
-
-        :param result: server response (dict)
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while deleting {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-        log.info("{} has been deleted".format(self.name()))
-        self.deleted_signal.emit()
-        self._module.removeNode(self)
-
     def update(self, new_settings):
         """
         Updates the settings for this VirtualBox VM.
@@ -227,68 +199,6 @@ class VirtualBoxVM(Node):
             log.info("VirtualBox VM {} has been updated".format(self.name()))
             self.updated_signal.emit()
 
-    def start(self):
-        """
-        Starts this VirtualBox VM instance.
-        """
-
-        if self.status() == Node.started:
-            log.debug("{} is already running".format(self.name()))
-            return
-
-        log.debug("{} is starting".format(self.name()))
-        self.httpPost("/virtualbox/vms/{vm_id}/start".format(vm_id=self._vm_id), self._startCallback)
-
-    def _startCallback(self, result, error=False, **kwargs):
-        """
-        Callback for start.
-
-        :param result: server response (dict)
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while starting {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-        else:
-            log.info("{} has started".format(self.name()))
-            self.setStatus(Node.started)
-            for port in self._ports:
-                # set ports as started
-                port.setStatus(Port.started)
-            self.started_signal.emit()
-
-    def stop(self):
-        """
-        Stops this VirtualBox VM instance.
-        """
-
-        if self.status() == Node.stopped:
-            log.debug("{} is already stopped".format(self.name()))
-            return
-
-        log.debug("{} is stopping".format(self.name()))
-        self.httpPost("/virtualbox/vms/{vm_id}/stop".format(vm_id=self._vm_id), self._stopCallback)
-
-    def _stopCallback(self, result, error=False, **kwargs):
-        """
-        Callback for stop.
-
-        :param result: server response (dict)
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while stopping {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-        else:
-            log.info("{} has stopped".format(self.name()))
-            self.setStatus(Node.stopped)
-            for port in self._ports:
-                # set ports as stopped
-                port.setStatus(Port.stopped)
-            self.stopped_signal.emit()
-
     def suspend(self):
         """
         Suspends this VirtualBox VM instance.
@@ -319,111 +229,6 @@ class VirtualBoxVM(Node):
                 # set ports as suspended
                 port.setStatus(Port.suspended)
             self.suspended_signal.emit()
-
-    def reload(self):
-        """
-        Reloads this VirtualBox VM instance.
-        """
-
-        log.debug("{} is being reloaded".format(self.name()))
-        self.httpPost("/virtualbox/vms/{vm_id}/reload".format(vm_id=self._vm_id), self._reloadCallback)
-
-    def _reloadCallback(self, result, error=False, **kwargs):
-        """
-        Callback for reload.
-
-        :param result: server response (dict)
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while reloading {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-        else:
-            log.info("{} has reloaded".format(self.name()))
-
-    def allocateUDPPort(self, port_id):
-        """
-        Requests an UDP port allocation.
-
-        :param port_id: port identifier
-        """
-
-        log.debug("{} is requesting an UDP port allocation".format(self.name()))
-        self._server.post("/ports/udp", partial(self._allocateUDPPortCallback, port_id))
-
-    def _allocateUDPPortCallback(self, port_id, result, error=False, **kwargs):
-        """
-        Callback for allocateUDPPort.
-
-        :param result: server response
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while allocating an UDP port for {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-        else:
-            lport = result["udp_port"]
-            log.debug("{} has allocated UDP port {}".format(self.name(), lport))
-            self.allocate_udp_nio_signal.emit(self.id(), port_id, lport)
-
-    def addNIO(self, port, nio):
-        """
-        Adds a new NIO on the specified port for this VirtualBox VM instance.
-
-        :param port: Port instance
-        :param nio: NIO instance
-        """
-
-        params = self.getNIOInfo(nio)
-        log.debug("{} is adding an {}: {}".format(self.name(), nio, params))
-
-        self.httpPost("/virtualbox/vms/{vm_id}/adapters/{adapter_id}/nio".format(vm_id=self._vm_id,
-                                                                                 adapter_id=port.portNumber()),
-                      partial(self._addNIOCallback, port.id()), body=params)
-
-    def _addNIOCallback(self, port_id, result, error=False, **kwargs):
-        """
-        Callback for addNIO.
-
-        :param result: server response
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while adding an UDP NIO for {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-            self.nio_cancel_signal.emit(self.id())
-        else:
-            self.nio_signal.emit(self.id(), port_id)
-
-    def deleteNIO(self, port):
-        """
-        Deletes an NIO from the specified port on this VirtualBox VM instance.
-
-        :param port: Port instance
-        """
-
-        log.debug("{} is deleting an NIO on port {}".format(self.name(), port.portNumber()))
-        self.httpDelete("/virtualbox/vms/{vm_id}/adapters/{adapter_id}/nio".format(vm_id=self._vm_id,
-                                                                                   adapter_id=port.portNumber()),
-                        self._deleteNIOCallback)
-
-    def _deleteNIOCallback(self, result, error=False, **kwargs):
-        """
-        Callback for deleteNIO.
-
-        :param result: server response
-        :param error: indicates an error (boolean)
-        """
-
-        if error:
-            log.error("error while deleting NIO {}: {}".format(self.name(), result["message"]))
-            self.server_error_signal.emit(self.id(), result["message"])
-            return
-
-        log.debug("{} has deleted a NIO: {}".format(self.name(), result))
 
     def startPacketCapture(self, port, capture_file_name, data_link_type):
         """
@@ -602,15 +407,6 @@ class VirtualBoxVM(Node):
         self._module.addNode(self)
         self._inital_settings = None
         self._loading = False
-
-    def vm_id(self):
-        """
-        Return the ID of this VirtualBox VM instance.
-
-        :returns: identifier (string)
-        """
-
-        return self._vm_id
 
     def name(self):
         """
