+++ tests/modules/virtualbox/test_virtualbox_vm.py
@@ -49,98 +49,6 @@ def test_virtualbox_vm_setup(virtualbox_vm, project):
         assert virtualbox_vm.vm_id() == "aec7a00c-e71c-45a6-8c04-29e40732883c"
 
 
-def test_virtualbox_vm_start(virtualbox_vm):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        virtualbox_vm.start()
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/virtualbox/vms/{vm_id}/start".format(vm_id=virtualbox_vm.vm_id())
-
-
-def test_virtualbox_vm_stop(virtualbox_vm):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        virtualbox_vm.setStatus(Node.started)
-        virtualbox_vm.stop()
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/virtualbox/vms/{vm_id}/stop".format(vm_id=virtualbox_vm.vm_id())
-
-
-def test_virtualbox_vm_reload(virtualbox_vm):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        virtualbox_vm.reload()
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/virtualbox/vms/{vm_id}/reload".format(vm_id=virtualbox_vm.vm_id())
-
-
-def test_allocateUDPPort(virtualbox_vm):
-
-    with patch('gns3.http_client.HTTPClient.post') as mock:
-        virtualbox_vm.allocateUDPPort(1)
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/ports/udp"
-
-        # Connect the signal
-        signal_mock = Mock()
-        virtualbox_vm.allocate_udp_nio_signal.connect(signal_mock)
-
-        # Callback
-        args[1]({"udp_port": 4242})
-
-        # Check the signal
-        assert signal_mock.called
-        args, kwargs = signal_mock.call_args
-        assert args[0] == virtualbox_vm.id()
-        assert args[1] == 1
-        assert args[2] == 4242
-
-
-def test_addNIO(virtualbox_vm):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        port = Port("Port 0")
-        port.setPortNumber(0)
-        nio = NIOUDP(4242, "127.0.0.1", 4243)
-        virtualbox_vm.addNIO(port, nio)
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/virtualbox/vms/{vm_id}/adapters/0/nio".format(vm_id=virtualbox_vm.vm_id())
-
-        # Connect the signal
-        signal_mock = Mock()
-        virtualbox_vm.nio_signal.connect(signal_mock)
-
-        # Callback
-        args[1]({})
-
-        # Check the signal
-        assert signal_mock.called
-        args, kwargs = signal_mock.call_args
-        assert args[0] == virtualbox_vm.id()
-        assert args[1] == port.id()
-
-
-def test_deleteNIO(virtualbox_vm):
-
-    with patch('gns3.node.Node.httpPost') as mock_post:
-        with patch('gns3.node.Node.httpDelete') as mock_delete:
-            port = Port("Port 0")
-            port.setPortNumber(0)
-            nio = NIOUDP(4242, "127.0.0.1", 4243)
-            virtualbox_vm.addNIO(port, nio)
-
-            virtualbox_vm.deleteNIO(port)
-            assert mock_delete.called
-
-            args, kwargs = mock_delete.call_args
-            assert args[0] == "/virtualbox/vms/{vm_id}/adapters/0/nio".format(vm_id=virtualbox_vm.vm_id())
-
-
 def test_update(virtualbox_vm):
 
     new_settings = {
