+++ tests/modules/vpcs/test_vpcs_device.py
@@ -118,96 +118,6 @@ def test_vpcs_device_setup_script_file(vpcs_device, project, tmpdir):
         assert vpcs_device.vm_id() == "aec7a00c-e71c-45a6-8c04-29e40732883c"
 
 
-def test_vpcs_device_start(vpcs_device):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        vpcs_device.start()
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/vpcs/vms/{vm_id}/start".format(vm_id=vpcs_device.vm_id())
-
-
-def test_vpcs_device_stop(vpcs_device):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        vpcs_device.setStatus(Node.started)
-        vpcs_device.stop()
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/vpcs/vms/{vm_id}/stop".format(vm_id=vpcs_device.vm_id())
-
-
-def test_vpcs_device_reload(vpcs_device):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        vpcs_device.reload()
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/vpcs/vms/{vm_id}/reload".format(vm_id=vpcs_device.vm_id())
-
-
-def test_allocateUDPPort(vpcs_device):
-
-    with patch('gns3.http_client.HTTPClient.post') as mock:
-        vpcs_device.allocateUDPPort(1)
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/ports/udp"
-
-        # Connect the signal
-        signal_mock = Mock()
-        vpcs_device.allocate_udp_nio_signal.connect(signal_mock)
-
-        # Callback
-        args[1]({"udp_port": 4242})
-
-        # Check the signal
-        assert signal_mock.called
-        args, kwargs = signal_mock.call_args
-        assert args[0] == vpcs_device.id()
-        assert args[1] == 1
-        assert args[2] == 4242
-
-
-def test_addNIO(vpcs_device):
-
-    with patch('gns3.node.Node.httpPost') as mock:
-        port = Port("Port 1")
-        nio = NIOUDP(4242, "127.0.0.1", 4243)
-        vpcs_device.addNIO(port, nio)
-        assert mock.called
-        args, kwargs = mock.call_args
-        assert args[0] == "/vpcs/vms/{vm_id}/ports/0/nio".format(vm_id=vpcs_device.vm_id())
-
-        # Connect the signal
-        signal_mock = Mock()
-        vpcs_device.nio_signal.connect(signal_mock)
-
-        # Callback
-        args[1]({})
-
-        # Check the signal
-        assert signal_mock.called
-        args, kwargs = signal_mock.call_args
-        assert args[0] == vpcs_device.id()
-        assert args[1] == port.id()
-
-
-def test_deleteNIO(vpcs_device):
-
-    with patch('gns3.node.Node.httpPost') as mock_post:
-        with patch('gns3.node.Node.httpDelete') as mock_delete:
-            port = Port("Port 1")
-            nio = NIOUDP(4242, "127.0.0.1", 4243)
-            vpcs_device.addNIO(port, nio)
-
-            vpcs_device.deleteNIO(port)
-            assert mock_delete.called
-
-            args, kwargs = mock_delete.call_args
-            assert args[0] == "/vpcs/vms/{vm_id}/ports/0/nio".format(vm_id=vpcs_device.vm_id())
-
-
 def test_exportConfig(tmpdir, vpcs_device):
 
     path = tmpdir / 'startup.vpcs'
